+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.2.0
 - Add --keep flag for preserving generated files
+- Replace patch inheritance with the --in-workspace flag
 # v0.1.2
 - Inherit patches when testing dependencies (https://github.com/wojciech-graj/cargo-no-std/pull/1)
 # v0.1.1